replace crio-wipe with crio wipe
before, crio-wipe would nuke the entire storage directory, not worrying about there being podman containers
instead, let's be more fine grained about wiping before an upgrade.

add a crio subcommand wipe, which wipes only crio storage.
update tests and docs accordingly
introduce package criocli that makes main code cleaner, and allows for sharing with wipe.go
