+++ internal/version/version_test.go
@@ -21,6 +21,37 @@ func fails(t *testing.T, err error) {
 	}
 }
 
+func mustUpgrade(t *testing.T, upgrade bool) {
+	if !upgrade {
+		t.Error("CRI-O should have upgraded")
+	}
+}
+
+func mustNotUpgrade(t *testing.T, upgrade bool) {
+	if upgrade {
+		t.Error("CRI-O should not have upgraded")
+	}
+}
+
+func createTempFile(t *testing.T) *os.File {
+	tmpFile, err := ioutil.TempFile("", "temporary-testing-file")
+	if err != nil {
+		t.Skip()
+	}
+	return tmpFile
+}
+
+func upgradeBetweenVersions(t *testing.T, oldVersion, newVersion string) (bool, error) {
+	tmpFile := createTempFile(t)
+	defer os.Remove(tmpFile.Name())
+
+	if err := writeVersionFile(tmpFile.Name(), "", oldVersion); err != nil {
+		return true, err
+	}
+
+	return shouldCrioWipe(tmpFile.Name(), newVersion)
+}
+
 func TestParseVersionCorrectVersion(t *testing.T) {
 	_, err := parseVersionConstant("1.1.1", "")
 	must(t, err)
@@ -28,7 +59,7 @@ func TestParseVersionCorrectVersion(t *testing.T) {
 	_, err = parseVersionConstant("1.1.1-dev", "")
 	must(t, err)
 
-	_, err = parseVersionConstant("1.1.1-dev", "bigloggitcommit")
+	_, err = parseVersionConstant("1.1.1-dev", "biglonggitcommit")
 	must(t, err)
 }
 
@@ -41,6 +72,7 @@ func TestParseVersionAddsGitCommit(t *testing.T) {
 	if len(v.Build) < 1 {
 		t.Error(errors.Errorf("Git commit not included in semver build"))
 	}
+
 	// git commit should have quotes removed
 	trimmed := strings.Trim(gitCommit, "\"")
 	if v.Build[0] != trimmed {
@@ -70,19 +102,18 @@ func TestParseVersionWithCurrentVersion(t *testing.T) {
 }
 
 func TestWriteVersionFileWritesFile(t *testing.T) {
-	tmpFile, err := ioutil.TempFile("", "temporary-testing-file")
-	if err != nil {
-		t.Skip()
-	}
+	tmpFile := createTempFile(t)
 	defer os.Remove(tmpFile.Name())
+
 	gitCommit := "fakeGitCommit"
-	err = WriteVersionFile(tmpFile.Name(), gitCommit)
+	version := "1.1.1"
+	err := writeVersionFile(tmpFile.Name(), gitCommit, version)
 	must(t, err)
 
 	versionBytes, err := ioutil.ReadFile(tmpFile.Name())
 	must(t, err)
 
-	versionConstantVersion, err := parseVersionConstant(Version, gitCommit)
+	versionConstantVersion, err := parseVersionConstant(version, gitCommit)
 	must(t, err)
 
 	versionConstantJSON, err := versionConstantVersion.MarshalJSON()
@@ -94,9 +125,67 @@ func TestWriteVersionFileWritesFile(t *testing.T) {
 
 func TestWriteVersionFileCreatesDir(t *testing.T) {
 	filename := "/tmp/crio/temporary-testing-file"
-	err := WriteVersionFile(filename, "")
+	err := writeVersionFile(filename, "", "1.1.1")
 	must(t, err)
 
 	_, err = ioutil.ReadFile(filename)
 	must(t, err)
+}
+
+func TestUpgradeWithUnspecifiedVersionFile(t *testing.T) {
+	upgrade, err := shouldCrioWipe("", "1.1.1")
+	mustUpgrade(t, upgrade)
+	fails(t, err)
+}
+
+func TestUpgradeWithEmptyVersionFile(t *testing.T) {
+	tmpFile := createTempFile(t)
+	defer os.Remove(tmpFile.Name())
+
+	upgrade, err := shouldCrioWipe(tmpFile.Name(), "1.1.1")
+	mustUpgrade(t, upgrade)
+	fails(t, err)
+}
+
+func TestFailUpgradeWithFaultyVersionFile(t *testing.T) {
+	tmpFile := createTempFile(t)
+	defer os.Remove(tmpFile.Name())
+	err := ioutil.WriteFile(tmpFile.Name(), []byte("bad version file"), 0644)
+	if err != nil {
+		t.Skip()
+	}
+
+	upgrade, err := shouldCrioWipe(tmpFile.Name(), "1.1.1")
+	mustUpgrade(t, upgrade)
+	fails(t, err)
+}
+
+func TestNoUpgradeWithSameVersion(t *testing.T) {
+	upgraded, err := upgradeBetweenVersions(t, "1.1.1", "1.1.1")
+	must(t, err)
+	mustNotUpgrade(t, upgraded)
+}
+
+func TestNoUpgradeWithSubMinorRelease(t *testing.T) {
+	upgraded, err := upgradeBetweenVersions(t, "1.1.1", "1.1.2")
+	must(t, err)
+	mustNotUpgrade(t, upgraded)
+}
+
+func TestUpgradeMinorRelease(t *testing.T) {
+	upgraded, err := upgradeBetweenVersions(t, "1.14.1", "1.13.1")
+	must(t, err)
+	mustUpgrade(t, upgraded)
+}
+
+func TestUpgradeMajorRelease(t *testing.T) {
+	upgraded, err := upgradeBetweenVersions(t, "2.0.0", "1.13.1")
+	must(t, err)
+	mustUpgrade(t, upgraded)
+}
+
+func TestFailNoUpgradeWithBadVersion(t *testing.T) {
+	upgraded, err := upgradeBetweenVersions(t, "bad version format", "1.13.1")
+	fails(t, err)
+	mustUpgrade(t, upgraded)
 }
